Adding security_token back and updating README
Adding security_token back in as it's needed for assumed roles. Also forcing the cloudwatch tasks run locally to be run without assuming root so the module can implicitly use env variables.

Found at that while using an assumed role the cloudwatch modules implicitly use the security_token from the env variables however due to the role requiring sudo via  it turned out that the env variables weren't there (for root) and so the module failed unless the security_token was set explicitly.