Reverting Already Installed due to bug
The package module returns different values depending on the package manage it uses. For centos, yum returns 'rc'. For ubuntu, apt returns only stdout and stderr.

I tried to rewrite this using conditionals instead of returning to the forced failed task (which is bad code) but the conditonal statements became to complex and didn't make any sense. I think there's a better solution but it's going to take too long to figure it out so I'm returning to the old method that was working dispite forcing the task to fail.